fixed ambiguous global export
