+++ src/rest_api/users/models.rs
@@ -1,17 +1,17 @@
 pub mod post_users_request;
-pub use self::post_users_request::*;
+pub use self::post_users_request::PostUsersRequest;
 
 pub mod post_users_terminate_sessions_request;
-pub use self::post_users_terminate_sessions_request::*;
+pub use self::post_users_terminate_sessions_request::PostUsersTerminateSessionsRequest;
 
 pub mod put_users_id_request;
 pub use self::put_users_id_request::*;
 
 pub mod put_users_id_request_notification_email;
-pub use self::put_users_id_request_notification_email::*;
+pub use self::put_users_id_request_notification_email::PutUsersIdRequestNotificationEmail;
 
 pub mod session_termination_message;
-pub use self::session_termination_message::*;
+pub use self::session_termination_message::SessionTerminationMessage;
 
 pub mod tracking_code;
 pub use self::tracking_code::TrackingCode;
@@ -30,3 +30,6 @@ pub use self::user_full_all_of_enterprise::UserFullAllOfEnterprise;
 
 pub mod users;
 pub use self::users::Users;
+
+pub mod users_common;
+pub use self::users_common::*;
