+++ src/rest_api/users/models/post_users_request.rs
@@ -8,7 +8,7 @@
  * Generated by: https://openapi-generator.tech
  */
 
-use super::tracking_code::TrackingCode;
+use super::{tracking_code::TrackingCode, Role, Status};
 
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct PostUsersRequest {
@@ -108,38 +108,5 @@ impl PostUsersRequest {
             status: None,
             external_app_user_id: None,
         }
-    }
-}
-
-/// The user's enterprise role
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub enum Role {
-    #[serde(rename = "coadmin")]
-    Coadmin,
-    #[serde(rename = "user")]
-    User,
-}
-
-impl Default for Role {
-    fn default() -> Role {
-        Self::Coadmin
-    }
-}
-/// The user's account status
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub enum Status {
-    #[serde(rename = "active")]
-    Active,
-    #[serde(rename = "inactive")]
-    Inactive,
-    #[serde(rename = "cannot_delete_edit")]
-    CannotDeleteEdit,
-    #[serde(rename = "cannot_delete_edit_upload")]
-    CannotDeleteEditUpload,
-}
-
-impl Default for Status {
-    fn default() -> Status {
-        Self::Active
     }
 }
