+++ src/rest_api/users/models/user_full.rs
@@ -8,7 +8,9 @@
  * Generated by: https://openapi-generator.tech
  */
 
-use super::{tracking_code::TrackingCode, user_full_all_of_enterprise::UserFullAllOfEnterprise};
+use super::{
+    tracking_code::TrackingCode, user_full_all_of_enterprise::UserFullAllOfEnterprise, Role, Status,
+};
 
 /// UserFull : A full representation of a user, as can be returned from any user API endpoint.
 
@@ -178,39 +180,5 @@ pub enum RHashType {
 impl Default for RHashType {
     fn default() -> RHashType {
         Self::User
-    }
-}
-/// The user's account status
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub enum Status {
-    #[serde(rename = "active")]
-    Active,
-    #[serde(rename = "inactive")]
-    Inactive,
-    #[serde(rename = "cannot_delete_edit")]
-    CannotDeleteEdit,
-    #[serde(rename = "cannot_delete_edit_upload")]
-    CannotDeleteEditUpload,
-}
-
-impl Default for Status {
-    fn default() -> Status {
-        Self::Active
-    }
-}
-/// The user’s enterprise role
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub enum Role {
-    #[serde(rename = "admin")]
-    Admin,
-    #[serde(rename = "coadmin")]
-    Coadmin,
-    #[serde(rename = "user")]
-    User,
-}
-
-impl Default for Role {
-    fn default() -> Role {
-        Self::Admin
     }
 }
